fix(server): others to validate only contain one object
